+++ mdm/api/src/main/java/com/gentics/mesh/core/data/node/field/list/HibHtmlFieldList.java
@@ -1,10 +1,17 @@
 package com.gentics.mesh.core.data.node.field.list;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
+import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.data.node.field.HibHtmlField;
 import com.gentics.mesh.core.data.node.field.nesting.HibMicroschemaListableField;
 import com.gentics.mesh.core.rest.node.field.list.impl.HtmlFieldListImpl;
+import com.gentics.mesh.util.CompareUtils;
 
 public interface HibHtmlFieldList extends HibMicroschemaListableField, HibListField<HibHtmlField, HtmlFieldListImpl, String> {
+
+	String TYPE = "html";
 
 	/**
 	 * Create a new html graph field.
@@ -22,4 +29,29 @@ public interface HibHtmlFieldList extends HibMicroschemaListableField, HibListFi
 	 */
 	HibHtmlField getHTML(int index);
 
+	@Override
+	default HtmlFieldListImpl transformToRest(InternalActionContext ac, String fieldKey, List<String> languageTags, int level) {
+		HtmlFieldListImpl restModel = new HtmlFieldListImpl();
+		for (HibHtmlField item : getList()) {
+			restModel.add(item.getHTML());
+		}
+		return restModel;
+	}
+
+	@Override
+	default List<String> getValues() {
+		return getList().stream().map(HibHtmlField::getHTML).collect(Collectors.toList());
+	}
+
+	@Override
+	default boolean listEquals(Object obj) {
+		if (obj instanceof HtmlFieldListImpl) {
+			HtmlFieldListImpl restField = (HtmlFieldListImpl) obj;
+			List<String> restList = restField.getItems();
+			List<? extends HibHtmlField> graphList = getList();
+			List<String> graphStringList = graphList.stream().map(e -> e.getHTML()).collect(Collectors.toList());
+			return CompareUtils.equals(restList, graphStringList);
+		}
+		return HibListField.super.listEquals(obj);
+	}
 }
