+++ mdm/api/src/main/java/com/gentics/mesh/core/data/node/field/list/HibNumberFieldList.java
@@ -1,10 +1,17 @@
 package com.gentics.mesh.core.data.node.field.list;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
+import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.data.node.field.HibNumberField;
 import com.gentics.mesh.core.data.node.field.nesting.HibMicroschemaListableField;
 import com.gentics.mesh.core.rest.node.field.list.impl.NumberFieldListImpl;
+import com.gentics.mesh.util.CompareUtils;
 
 public interface HibNumberFieldList extends HibMicroschemaListableField, HibListField<HibNumberField, NumberFieldListImpl, Number> {
+	String TYPE = "number";
+
 	/**
 	 * Create a new number graph field with the given value.
 	 * 
@@ -21,4 +28,29 @@ public interface HibNumberFieldList extends HibMicroschemaListableField, HibList
 	 */
 	HibNumberField getNumber(int index);
 
+	@Override
+	default NumberFieldListImpl transformToRest(InternalActionContext ac, String fieldKey, List<String> languageTags, int level) {
+		NumberFieldListImpl restModel = new NumberFieldListImpl();
+		for (HibNumberField item : getList()) {
+			restModel.add(item.getNumber());
+		}
+		return restModel;
+	}
+
+	@Override
+	default List<Number> getValues() {
+		return getList().stream().map(HibNumberField::getNumber).collect(Collectors.toList());
+	}
+
+	@Override
+	default boolean listEquals(Object obj) {
+		if (obj instanceof NumberFieldListImpl) {
+			NumberFieldListImpl restField = (NumberFieldListImpl) obj;
+			List<Number> restList = restField.getItems();
+			List<? extends HibNumberField> graphList = getList();
+			List<Number> graphStringList = graphList.stream().map(e -> e.getNumber()).collect(Collectors.toList());
+			return CompareUtils.equals(restList, graphStringList);
+		}
+		return HibListField.super.listEquals(obj);
+	}
 }
