+++ mdm/orientdb-api/src/main/java/com/gentics/mesh/core/data/node/field/list/NodeGraphFieldList.java
@@ -4,16 +4,9 @@ import com.gentics.mesh.core.data.node.HibNode;
 import com.gentics.mesh.core.data.node.field.nesting.HibNodeField;
 import com.gentics.mesh.core.rest.node.field.list.NodeFieldList;
 
-import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.LoggerFactory;
-
 /**
  * List definition for node list field.
  */
 public interface NodeGraphFieldList extends ListGraphField<HibNodeField, NodeFieldList, HibNode>, HibNodeFieldList {
-
-	Logger log = LoggerFactory.getLogger(NodeGraphFieldList.class);
-
-	String TYPE = "node";
 
 }
