+++ mdm/orientdb-wrapper/src/main/java/com/gentics/mesh/core/data/node/field/list/impl/DateGraphFieldListImpl.java
@@ -1,14 +1,8 @@
 package com.gentics.mesh.core.data.node.field.list.impl;
 
-import static com.gentics.mesh.util.DateUtils.toISO8601;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
 import com.gentics.madl.index.IndexHandler;
 import com.gentics.madl.type.TypeHandler;
 import com.gentics.mesh.context.BulkActionContext;
-import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.data.generic.MeshVertexImpl;
 import com.gentics.mesh.core.data.node.field.DateGraphField;
 import com.gentics.mesh.core.data.node.field.HibDateField;
@@ -16,7 +10,6 @@ import com.gentics.mesh.core.data.node.field.impl.DateGraphFieldImpl;
 import com.gentics.mesh.core.data.node.field.list.AbstractBasicGraphFieldList;
 import com.gentics.mesh.core.data.node.field.list.DateGraphFieldList;
 import com.gentics.mesh.core.rest.node.field.list.impl.DateFieldListImpl;
-import com.gentics.mesh.util.CompareUtils;
 
 /**
  * @see DateGraphFieldList
@@ -59,31 +52,4 @@ public class DateGraphFieldListImpl extends AbstractBasicGraphFieldList<HibDateF
 	public void delete(BulkActionContext bac) {
 		getElement().remove();
 	}
-
-	@Override
-	public DateFieldListImpl transformToRest(InternalActionContext ac, String fieldKey, List<String> languageTags, int level) {
-		DateFieldListImpl restModel = new DateFieldListImpl();
-		for (HibDateField item : getList()) {
-			restModel.add(toISO8601(item.getDate()));
-		}
-		return restModel;
-	}
-
-	@Override
-	public List<Long> getValues() {
-		return getList().stream().map(HibDateField::getDate).collect(Collectors.toList());
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (obj instanceof DateFieldListImpl) {
-			DateFieldListImpl restField = (DateFieldListImpl) obj;
-			List<String> restList = restField.getItems();
-			List<? extends HibDateField> graphList = getList();
-			List<String> graphStringList = graphList.stream().map(e -> toISO8601(e.getDate())).collect(Collectors.toList());
-			return CompareUtils.equals(restList, graphStringList);
-		}
-		return super.equals(obj);
-	}
-
 }
