+++ mdm/orientdb-wrapper/src/main/java/com/gentics/mesh/core/data/node/field/list/impl/HtmlGraphFieldListImpl.java
@@ -1,12 +1,8 @@
 package com.gentics.mesh.core.data.node.field.list.impl;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 import com.gentics.madl.index.IndexHandler;
 import com.gentics.madl.type.TypeHandler;
 import com.gentics.mesh.context.BulkActionContext;
-import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.data.generic.MeshVertexImpl;
 import com.gentics.mesh.core.data.node.field.HibHtmlField;
 import com.gentics.mesh.core.data.node.field.HtmlGraphField;
@@ -14,7 +10,6 @@ import com.gentics.mesh.core.data.node.field.impl.HtmlGraphFieldImpl;
 import com.gentics.mesh.core.data.node.field.list.AbstractBasicGraphFieldList;
 import com.gentics.mesh.core.data.node.field.list.HtmlGraphFieldList;
 import com.gentics.mesh.core.rest.node.field.list.impl.HtmlFieldListImpl;
-import com.gentics.mesh.util.CompareUtils;
 
 /**
  * @see HtmlGraphFieldList
@@ -56,31 +51,5 @@ public class HtmlGraphFieldListImpl extends AbstractBasicGraphFieldList<HibHtmlF
 	@Override
 	public void delete(BulkActionContext bac) {
 		getElement().remove();
-	}
-
-	@Override
-	public HtmlFieldListImpl transformToRest(InternalActionContext ac, String fieldKey, List<String> languageTags, int level) {
-		HtmlFieldListImpl restModel = new HtmlFieldListImpl();
-		for (HibHtmlField item : getList()) {
-			restModel.add(item.getHTML());
-		}
-		return restModel;
-	}
-
-	@Override
-	public List<String> getValues() {
-		return getList().stream().map(HibHtmlField::getHTML).collect(Collectors.toList());
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (obj instanceof HtmlFieldListImpl) {
-			HtmlFieldListImpl restField = (HtmlFieldListImpl) obj;
-			List<String> restList = restField.getItems();
-			List<? extends HibHtmlField> graphList = getList();
-			List<String> graphStringList = graphList.stream().map(e -> e.getHTML()).collect(Collectors.toList());
-			return CompareUtils.equals(restList, graphStringList);
-		}
-		return super.equals(obj);
 	}
 }
