+++ mdm/orientdb-wrapper/src/main/java/com/gentics/mesh/core/data/node/field/list/impl/StringGraphFieldListImpl.java
@@ -1,28 +1,15 @@
 package com.gentics.mesh.core.data.node.field.list.impl;
 
-import static com.gentics.mesh.core.rest.error.Errors.error;
-import static io.netty.handler.codec.http.HttpResponseStatus.BAD_REQUEST;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
 import com.gentics.madl.index.IndexHandler;
 import com.gentics.madl.type.TypeHandler;
 import com.gentics.mesh.context.BulkActionContext;
-import com.gentics.mesh.context.InternalActionContext;
-import com.gentics.mesh.core.data.HibField;
 import com.gentics.mesh.core.data.generic.MeshVertexImpl;
-import com.gentics.mesh.core.data.node.field.FieldGetter;
-import com.gentics.mesh.core.data.node.field.FieldTransformer;
-import com.gentics.mesh.core.data.node.field.FieldUpdater;
 import com.gentics.mesh.core.data.node.field.HibStringField;
 import com.gentics.mesh.core.data.node.field.StringGraphField;
 import com.gentics.mesh.core.data.node.field.impl.StringGraphFieldImpl;
 import com.gentics.mesh.core.data.node.field.list.AbstractBasicGraphFieldList;
-import com.gentics.mesh.core.data.node.field.list.HibStringFieldList;
 import com.gentics.mesh.core.data.node.field.list.StringGraphFieldList;
 import com.gentics.mesh.core.rest.node.field.list.impl.StringFieldListImpl;
-import com.gentics.mesh.util.CompareUtils;
 
 /**
  * @see StringGraphFieldList
@@ -66,31 +53,4 @@ public class StringGraphFieldListImpl extends AbstractBasicGraphFieldList<HibStr
 	public void delete(BulkActionContext bac) {
 		getElement().remove();
 	}
-
-	@Override
-	public StringFieldListImpl transformToRest(InternalActionContext ac, String fieldKey, List<String> languageTags, int level) {
-		StringFieldListImpl restModel = new StringFieldListImpl();
-		for (HibStringField item : getList()) {
-			restModel.add(item.getString());
-		}
-		return restModel;
-	}
-
-	@Override
-	public List<String> getValues() {
-		return getList().stream().map(HibStringField::getString).collect(Collectors.toList());
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (obj instanceof StringFieldListImpl) {
-			StringFieldListImpl restField = (StringFieldListImpl) obj;
-			List<String> restList = restField.getItems();
-			List<? extends HibStringField> graphList = getList();
-			List<String> graphStringList = graphList.stream().map(e -> e.getString()).collect(Collectors.toList());
-			return CompareUtils.equals(restList, graphStringList);
-		}
-		return super.equals(obj);
-	}
-
 }
